+++ QCamera2/HAL/test/qcamera_test.h
@@ -55,7 +55,7 @@ public:
     typedef struct {
         unsigned char *  Data;
         int      Type;
-        size_t   Size;
+        unsigned Size;
     } Sections_t;
 
 public:
@@ -96,7 +96,7 @@ public:
     status_t unconfigureRecorder();
     Sections_t *FindSection(int SectionType);
     status_t ReadSectionsFromBuffer (unsigned char *buffer,
-            size_t buffer_size, ReadMode_t ReadMode);
+        unsigned int buffer_size, ReadMode_t ReadMode);
     virtual IBinder* onAsBinder();
     void setTestCtxInstance(TestContext *instance);
 
@@ -134,7 +134,9 @@ public:
 
 private:
 
-    status_t createPreviewSurface(int width, int height, int32_t pixFormat);
+    status_t createPreviewSurface(unsigned int width,
+            unsigned int height,
+            int32_t pixFormat);
     status_t destroyPreviewSurface();
 
     status_t saveFile(const sp<IMemory>& mem, String8 path);
@@ -156,11 +158,11 @@ private:
     bool mDoPrintMenu;
     bool mPiPCapture;
     static int mPiPIdx;
-    unsigned int mfmtMultiplier;
+    int mfmtMultiplier;
     int mWidthTmp;
     int mHeightTmp;
-    size_t mSectionsRead;
-    size_t mSectionsAllocated;
+    int mSectionsRead;
+    int mSectionsAllocated;
     Sections_t * mSections;
     Sections_t * mJEXIFTmp;
     Sections_t mJEXIFSection;
@@ -178,9 +180,9 @@ private:
     SkBitmap skBMtmp;
     sp<IMemory> PiPPtrTmp;
 
-    size_t mCurrentPreviewSizeIdx;
-    size_t mCurrentPictureSizeIdx;
-    size_t mCurrentVideoSizeIdx;
+    int mCurrentPreviewSizeIdx;
+    int mCurrentPictureSizeIdx;
+    int mCurrentVideoSizeIdx;
     Vector<Size> mSupportedPreviewSizes;
     Vector<Size> mSupportedPictureSizes;
     Vector<Size> mSupportedVideoSizes;
@@ -283,7 +285,7 @@ private:
     static const int numberOfCommands;
 
     bool mUseScript;
-    size_t mCmdIndex;
+    int mCmdIndex;
     char *mScript;
     Vector<Command> mCommands;
     TestContext *mTestContext;
@@ -298,7 +300,7 @@ public:
     TestContext();
     ~TestContext();
 
-    size_t GetCamerasNum();
+    int32_t GetCamerasNum();
     status_t FunctionalTest();
     status_t AddScriptFromFile(const char *scriptFile);
     void setViVSize(Size VideoSize, int camIndex);
@@ -310,8 +312,8 @@ public:
 private:
     sp<CameraContext> camera[MAX_CAM_INSTANCES];
     char GetNextCmd(sp<qcamera::CameraContext> currentCamera);
-    size_t mCurrentCameraIndex;
-    size_t mSaveCurrentCameraIndex;
+    int  mCurrentCameraIndex;
+    int  mSaveCurrentCameraIndex;
     Vector< sp<qcamera::CameraContext> > mAvailableCameras;
     bool mTestRunning;
     Interpreter *mInterpreter;
